-Wextra

+++ orter/bbc.c
@@ -62,7 +62,7 @@ static int orter_bbc_uef_write(char *name, uint16_t load, uint16_t exec)
 
   while (block_nr * 256 < (int) s) {
     uint16_t i = block_nr * 256;
-    uint16_t j = MIN(i + 256, s);
+    uint16_t j = MIN((size_t) i + 256, s);
     uint8_t *block = &data[i];
 
     /* construct data header */
